Added Player SeekBar, adjusted to Util changes

+++ src/main/xerus/monstercat/MonsterUtilities.kt
@@ -17,6 +17,7 @@ import xerus.ktutil.javafx.applySkin
 import xerus.ktutil.javafx.fill
 import xerus.ktutil.javafx.launch
 import xerus.ktutil.javafx.onJFX
+import xerus.ktutil.javafx.properties.listen
 import xerus.ktutil.javafx.ui.App
 import xerus.ktutil.javafx.ui.Changelog
 import xerus.ktutil.javafx.ui.JFXMessageDisplay
@@ -133,28 +134,35 @@ class MonsterUtilities : VBox(), JFXMessageDisplay {
         addTab(TabDownloader::class)
         addTab(TabSettings::class)
         if (VERSION != Settings.LASTVERSION.get()) {
-            launch {
-                logger.fine("New version detected! $VERSION from " + Settings.LASTVERSION())
-                val f = Settings.DELETE()
-                if (f.exists()) {
-                    logger.config("Deleting older version $f...")
-                    val time = currentSeconds()
-                    var res = false
-                    do {
-                        res = f.delete()
-                    } while (!res && time + 10 > currentSeconds())
-                    if (res) {
-                        Settings.DELETE.reset()
-                        logger.config("Deleted $f!")
-                    } else
-                        logger.config("Couldn't delete older version $f")
+            if (Settings.LASTVERSION().isEmpty()) {
+                logger.info("First launch! Showing tutorial!")
+                onJFX {
+                    // TODO intro dialog
                 }
-                Settings.LASTVERSION.put(VERSION)
+            } else {
+                launch {
+                    logger.fine("New version detected! $VERSION from " + Settings.LASTVERSION())
+                    val f = Settings.DELETE()
+                    if (f.exists()) {
+                        logger.config("Deleting older version $f...")
+                        val time = currentSeconds()
+                        var res = false
+                        do {
+                            res = f.delete()
+                        } while (!res && time + 10 > currentSeconds())
+                        if (res) {
+                            Settings.DELETE.reset()
+                            logger.config("Deleted $f!")
+                        } else
+                            logger.config("Couldn't delete older version $f")
+                    }
+                    Settings.LASTVERSION.put(VERSION)
+                }
+                showChangelog()
             }
-            showChangelog()
         }
 
-        children.add(Player)
+        children.add(Player.box)
         fill(tabPane)
         checkForUpdate()
     }
@@ -177,7 +185,7 @@ class MonsterUtilities : VBox(), JFXMessageDisplay {
                 else
                     onJFX {
                         val dialog = showAlert(Alert.AlertType.CONFIRMATION, "Updater", null, "New version $latestVersion available! Update now?", ButtonType.YES, ButtonType("Not now", ButtonBar.ButtonData.NO), ButtonType("Ignore this update", ButtonBar.ButtonData.CANCEL_CLOSE))
-                        dialog.resultProperty().addListener { _, _, type ->
+                        dialog.resultProperty().listen { type ->
                             if (type.buttonData == ButtonBar.ButtonData.YES) {
                                 update(latestVersion)
                             } else if (type.buttonData == ButtonBar.ButtonData.CANCEL_CLOSE)
@@ -232,11 +240,13 @@ class MonsterUtilities : VBox(), JFXMessageDisplay {
 
     fun showChangelog() {
         val c = Changelog("Note: The Catalog and Genres Tab pull their data from the MCatalog Spreadsheet, thus issues may stem from their side.").apply {
-            version(1, 0, "Release", "Brand new shiny favicon and player buttons - big thanks to NocFA!", "Feedback can now be sent directly from the application!")
+            version(1, 0, "Release", "Brand new shiny favicon and player buttons - big thanks to NocFA!",
+                    "Added tutorial", "Feedback can now be sent directly from the application!")
                     .change("New Downloader!", "Can download any combinations of Releases and Tracks", "Easy filtering", "connect.sid is now checked live", "Two distinct filename patterns for Singles and Album tracks", "Greatly improved pattern syntax with higher flexibility")
                     .change("Settings reworked", "Multiple skins available, changeable on-the-fly", "Startup Tab can now also be the previously opened one")
                     .change("Catalog and Genre Tab now show Genre colors")
                     .change("Catalog improved", "More filtering options", "Smart column size")
+                    .change("Player now has a Seekbar")
 
             /*
             version(0, 3, "UI Rework started", "Genres are now presented as a tree",
